+++ scripts/nanopineoimage.sh
@@ -1,6 +1,6 @@
 #!/bin/sh
 
-# Build Architecture Debian 32bit (to be changed to armv8)
+# Build Architecture Debian 32bit
 ARCH="armv7"
 
 while getopts ":v:p:a:" opt; do
@@ -18,7 +18,7 @@ while getopts ":v:p:a:" opt; do
 done
 
 BUILDDATE=$(date -I)
-IMG_FILE="Volumio${VERSION}-${BUILDDATE}-sopine64.img"
+IMG_FILE="Volumio${VERSION}-${BUILDDATE}-${ARCH}-nanopineo.img"
 
 if [ "$ARCH" = arm ]; then
   DISTRO="Raspbian"
@@ -60,24 +60,25 @@ sudo mkfs -F -t ext4 -L volumio "${SYS_PART}"
 sudo mkfs -F -t ext4 -L volumio_data "${DATA_PART}"
 sync
 
-echo "Preparing for the (so)Pine64(LTS) kernel/ platform files"
-if [ -d platform-pine64 ]
+echo "Preparing for the nanopineo kernel/ platform files"
+if [ -d platform-nanopineo ]
 then
 	echo "Platform folder already exists - keeping it"
-    # if you really want to re-clone from the repo, then delete the platform-pine64 folder
+    # if you really want to re-clone from the repo, then delete the platform-nanopineo folder
     # that will refresh all the odroid platforms, see below
 else
-	echo "Clone (so)Pine64(LTS) files from repo"
-	git clone https://github.com/volumio/platform-pine64.git platform-pine64
-	echo "Unpack the (so)Pine64(LTS) platform files"
-    cd platform-pine64
-	tar xfJ sopine64lts.tar.xz
+	echo "Clone nanopineo files from repo"
+	sudo mkdir platform-nanopineo
+    cd platform-nanopineo
+	wget https://github.com/nikkov/nanopineo2-platform/raw/master/nanopineo.tar.xz
+	echo "Unpack the platform files"
+  	tar xfJ nanopineo.tar.xz
+	rm nanopineo.tar.xz
 	cd ..
 fi
 
-echo "Copying the soPine64 (and Pine64LTS) bootloader"
-sudo dd if=platform-pine64/sopine64/u-boot/boot0-pine64-sopine.bin of=${LOOP_DEV} conv=notrunc bs=1k seek=8
-sudo dd if=platform-pine64/sopine64/u-boot/u-boot-pine64-sopine.bin of=${LOOP_DEV} conv=notrunc bs=1k seek=19096
+echo "Copying the bootloader"
+sudo dd if=platform-nanopineo/nanopineo/u-boot/u-boot-sunxi-with-spl.bin of=${LOOP_DEV} bs=1024 seek=8
 sync
 
 echo "Preparing for Volumio rootfs"
@@ -105,25 +106,27 @@ sudo mount -t vfat "${BOOT_PART}" /mnt/volumio/rootfs/boot
 
 echo "Copying Volumio RootFs"
 sudo cp -pdR build/$ARCH/root/* /mnt/volumio/rootfs
-echo "Copying (so)Pine64(LTS) boot files"
-mkdir /mnt/volumio/rootfs/boot/pine64
-sudo cp platform-pine64/sopine64/boot/pine64/Image /mnt/volumio/rootfs/boot/pine64
-sudo cp platform-pine64/sopine64/boot/pine64/*.dtb /mnt/volumio/rootfs/boot/pine64
-sudo cp platform-pine64/sopine64/boot/uEnv.txt /mnt/volumio/rootfs/boot
-sudo cp platform-pine64/sopine64/boot/Image.version /mnt/volumio/rootfs/boot
-sudo cp platform-pine64/sopine64/boot/config* /mnt/volumio/rootfs/boot
+echo "Copying nanopineo boot files"
+sudo cp platform-nanopineo/nanopineo/boot/zImage /mnt/volumio/rootfs/boot
+sudo cp platform-nanopineo/nanopineo/boot/*.dtb /mnt/volumio/rootfs/boot
 
-echo "Copying (so)Pine64(LTS) modules and firmware"
-sudo cp -pdR platform-pine64/sopine64/lib/modules /mnt/volumio/rootfs/lib/
-sudo cp -pdR platform-pine64/sopine64/lib/firmware /mnt/volumio/rootfs/lib/
+sudo cp platform-nanopineo/nanopineo/boot/boot.cmd /mnt/volumio/rootfs/boot
+sudo cp platform-nanopineo/nanopineo/boot/boot.scr /mnt/volumio/rootfs/boot
+
+sudo cp platform-nanopineo/nanopineo/boot/zImage.version /mnt/volumio/rootfs/boot
+sudo cp platform-nanopineo/nanopineo/boot/config* /mnt/volumio/rootfs/boot
+
+echo "Copying nanopineo modules and firmware"
+sudo cp -pdR platform-nanopineo/nanopineo/lib/modules /mnt/volumio/rootfs/lib/
+sudo cp -pdR platform-nanopineo/nanopineo/lib/firmware /mnt/volumio/rootfs/lib/
 
 echo "Confguring ALSA with sane defaults"
-sudo cp platform-pine64/sopine64/var/lib/alsa/* /mnt/volumio/rootfs/var/lib/alsa
+#sudo cp platform-nanopineo2/nanopineo/var/lib/alsa/* /mnt/volumio/rootfs/var/lib/alsa
 
 sync
 
-echo "Preparing to run chroot for more SOPINE A64 configuration (equals pine64)"
-cp scripts/pine64config.sh /mnt/volumio/rootfs
+echo "Preparing to run chroot for more nanopineo configuration"
+cp scripts/nanopineoconfig.sh /mnt/volumio/rootfs
 cp scripts/initramfs/init /mnt/volumio/rootfs/root
 cp scripts/initramfs/mkinitramfs-custom.sh /mnt/volumio/rootfs/usr/local/sbin
 #copy the scripts for updating from usb
@@ -136,27 +139,22 @@ echo $PATCH > /mnt/volumio/rootfs/patch
 
 chroot /mnt/volumio/rootfs /bin/bash -x <<'EOF'
 su -
-/pine64config.sh
+/nanopineoconfig.sh
 EOF
 
 #cleanup
-rm /mnt/volumio/rootfs/pine64config.sh /mnt/volumio/rootfs/root/init
+rm /mnt/volumio/rootfs/nanopineoconfig.sh /mnt/volumio/rootfs/root/init
 
 echo "Unmounting Temp devices"
 umount -l /mnt/volumio/rootfs/dev
 umount -l /mnt/volumio/rootfs/proc
 umount -l /mnt/volumio/rootfs/sys
 
-#echo "Copying LIRC configuration files"
-#sudo cp platform-pine64/pine64/etc/lirc/lircd.conf /mnt/volumio/rootfs/etc/lirc
-#sudo cp platform-pine64/pine64/etc/lirc/hardware.conf /mnt/volumio/rootfs/etc/lirc
-#sudo cp platform-pine64/pine64/etc/lirc/lircrc /mnt/volumio/rootfs/etc/lirc
-
-echo "==> soPine64/ Pine64 LTS device installed"
+echo "==> nanopineo device installed"
 
 #echo "Removing temporary platform files"
 #echo "(you can keep it safely as long as you're sure of no changes)"
-#sudo rm -r platform-pine64
+#sudo rm -r platform-nanopineo
 sync
 
 echo "Preparing rootfs base for SquashFS"
@@ -172,19 +170,24 @@ fi
 echo "Copying Volumio rootfs to Temp Dir"
 cp -rp /mnt/volumio/rootfs/* /mnt/squash/
 
-if [ -e /mnt/kernel_current.tar ]; then
+if [ -e /mnt/kernel_current.tar.gz ]; then
 	echo "Volumio Kernel Partition Archive exists - Cleaning it"
-	rm -rf /mnt/kernel_current.tar
+	rm -rf /mnt/kernel_current.tar.gz
 fi
 
 echo "Creating Kernel Partition Archive"
-tar cf /mnt/kernel_current.tar --exclude='resize-volumio-datapart' -C /mnt/squash/boot/ .
+tar cf /mnt/kernel_current.tar.gz --exclude='resize-volumio-datapart' -C /mnt/squash/boot/ .
 
 echo "Removing the Kernel"
 rm -rf /mnt/squash/boot/*
 
 echo "Creating SquashFS, removing any previous one"
-rm -r Volumio.sqsh
+if [ -e Volumio.sqsh ]; then
+	echo "Volumio Kernel Partition Archive exists - Cleaning it"
+	rm -r Volumio.sqsh
+fi
+
+echo "Creating SquashFS"
 mksquashfs /mnt/squash/* Volumio.sqsh
 
 echo "Squash filesystem created"
